kahee/requirements.txt
# LLM & 벡터 스택
chromadb==0.4.24   # 프로젝트 코드 전체 0.4.24 기준으로 코드 작성
langchain==0.2.1
langchain-community==0.2.1
langchain-openai==0.1.7
openai==1.78.1
sentence-transformers==4.1.0

# FastAPI 서버
fastapi==0.115.12
uvicorn[standard]==0.29.0

# 설정 & 환경
python-dotenv==1.0.1
pydantic==2.11.4
pydantic-settings==2.9.1 

# 데이터 처리
numpy==1.26.4    # chromadb-0.4.x는 NumPy 2.0 사용하면 충돌 -> 버전 고정
pandas==2.2.2
tqdm==4.66.2
requests==2.31.0

# 프론트
streamlit==1.34.0